Relocate Calculator component
