+++ src/components/App.js
@@ -1,8 +1,7 @@
-/* eslint-disable */
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import '../styles/App.css';
-import Calculator from '../routes/Calculator';
+import Calculator from './Calculator';
 import Header from './Header';
 import Home from '../routes/Home';
 import Quote from '../routes/Quote';
